+++ src/sdhlt/sdHLBSP/bsp5.h
@@ -45,6 +45,7 @@
 #define DEFAULT_NOBRINK			false
 #define DEFAULT_NOCLIP          false
 #define DEFAULT_NOOPT			false
+#define DEFAULT_LMOPTIMIZE		false
 #define DEFAULT_NOCLIPNODEMERGE	false
 #define DEFAULT_LEAKONLY        false
 #define DEFAULT_WATERVIS        false
@@ -289,6 +290,7 @@ extern bool		g_noinsidefill;
 extern bool     g_notjunc;
 extern bool		g_nobrink;
 extern bool		g_noclipnodemerge;
+extern bool		g_lmoptimize;
 extern bool     g_watervis;
 extern bool     g_chart;
 extern bool     g_estimate;