+++ src/sdhlt/sdHLBSP/qbsp.cpp
@@ -56,6 +56,7 @@ char			g_extentfilename[_MAX_PATH];
 
 // command line flags
 bool			g_noopt = DEFAULT_NOOPT;		// don't optimize BSP on write
+bool			g_lmoptimize = DEFAULT_LMOPTIMIZE;	// reorder faces to waste fewer lightmap atlas pages
 bool			g_noclipnodemerge = DEFAULT_NOCLIPNODEMERGE;
 bool            g_nofill = DEFAULT_NOFILL;      // dont fill "-nofill"
 bool			g_noinsidefill = DEFAULT_NOINSIDEFILL;
@@ -1389,6 +1390,7 @@ static void     Usage()
     Log("    -nofill        : Don't fill outside (will mask LEAKs) (not for final runs)\n");
 	Log("    -noinsidefill  : Don't fill empty spaces\n");
 	Log("    -noopt         : Don't optimize planes on BSP write   (not for final runs)\n");
+	Log("    -lmoptimize    : Reorder faces to waste fewer lightmap atlas pages\n");
 	Log("    -noclipnodemerge: Don't optimize clipnodes\n");
     Log("    -texdata #     : Alter maximum texture memory limit (in kb)\n");
     Log("    -lightdata #   : Alter maximum lighting memory limit (in kb)\n");
@@ -1470,6 +1472,7 @@ static void     Settings()
 	Log("noinsidefill        [ %7s ] [ %7s ]\n", g_noinsidefill ? "on" : "off", DEFAULT_NOINSIDEFILL ? "on" : "off");
 	Log("noopt               [ %7s ] [ %7s ]\n", g_noopt ? "on" : "off", DEFAULT_NOOPT ? "on" : "off");
 	Log("no clipnode merging [ %7s ] [ %7s ]\n", g_noclipnodemerge? "on": "off", DEFAULT_NOCLIPNODEMERGE? "on": "off");
+	Log("lightmap atlas opt. [ %7s ] [ %7s ]\n", g_lmoptimize? "on": "off", DEFAULT_LMOPTIMIZE? "on": "off");
     Log("null tex. stripping [ %7s ] [ %7s ]\n", g_bUseNullTex ? "on" : "off", DEFAULT_NULLTEX ? "on" : "off" );
     Log("notjunc             [ %7s ] [ %7s ]\n", g_notjunc ? "on" : "off", DEFAULT_NOTJUNC ? "on" : "off");
 	Log("nobrink             [ %7s ] [ %7s ]\n", g_nobrink? "on": "off", DEFAULT_NOBRINK? "on": "off");
@@ -1770,6 +1773,10 @@ int             main(const int argc, char** argv)
 		else if (!strcasecmp(argv[i], "-noopt"))
 		{
 			g_noopt = true;
+		}
+		else if (!strcasecmp(argv[i], "-lmoptimize"))
+		{
+			g_lmoptimize = true;
 		}
 		else if (!strcasecmp (argv[i], "-noclipnodemerge"))
 		{